Add files via upload
This project is to use HTML, CSS and JavaScript to create a working to do list. Note
that a to do list is an HTML unordered list (<ul>) with CSS formatting and JavaScript
functionality. Your goal is to create a well-designed user interface using client-side
programming. You do not need to store any information on a server or in a database.